climbing-scrape/requirements.txt
boto3
bs4
pyyaml
requests
